MVC 패턴에 맞게 폴더 생성
