--- kickBoardProject/ViewController.swift
@@ -1,19 +1,0 @@
-//
-//  ViewController.swift
-//  kickBoardProject
-//
-//  Created by 박승환 on 7/23/24.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
-
