+++ kickBoardProject/controller/ViewController.swift
@@ -1,0 +1,11 @@
+import UIKit
+
+class ViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        
+    }
+
+
+}
